LayoutTests/http/tests/paymentrequest/payment-request-change-shipping-option.https-expected.txt

CONSOLE MESSAGE: WebKit currently uses the first shipping option even if other shipping options are marked as selected.

FAIL Test for PaymentRequest shippingOption attribute assert_equals: Must be 'initially-selected', as the selected member is true expected "initially-selected" but got "valid-1"
